requirements.txt
# Web framework
fastapi>=0.115
uvicorn[standard]>=0.32

# Database
aiosqlite>=0.20

# LiveKit voice pipeline
livekit>=0.17
livekit-agents>=0.12
livekit-plugins-openai>=0.3
livekit-plugins-deepgram>=0.7
livekit-plugins-cartesia>=0.4
livekit-plugins-silero>=0.7

# OpenAI
openai>=1.55

# Utils
python-dotenv>=1.0
pydantic>=2.10
httpx>=0.27
